Added app:reset rake tas.

+++ db/migrate/20140912072904_add_topic_id_to_posts.rb
@@ -1,5 +1,5 @@
 class AddTopicIdToPosts < ActiveRecord::Migration
   def change
-    add_column :posts, :topic_id, :integer, unique: true
+    add_column :posts, :topic_id, :bigint, unique: true
   end
 end
